Moved QueryNodes into own classes
