+++ cql/src/test/java/dk/dbc/opensearch/cql/CQLParserTest.java
@@ -68,7 +68,7 @@ public class CQLParserTest {
     public void testCase() throws Exception {
         System.out.println(name);
         try {
-            Query result = new CQLParser(query, VALIDATORS).parse();
+            QueryNode result = new CQLParser(query, VALIDATORS).parse();
             String actual = result.toString();
             System.out.println("actual = " + actual);
             System.out.println("expected = " + expected);
